Refactored phase1 phis
Keeping old phase1 phis (phis_assemble_example) as a depricated example for using assembly.
